requirements.txt
openpyxl
pandas
gensim
textblob
boto3
python-dotenv
requests
praw